+ name & license
Sorry, I should have seen this before.  Helpful to web search author name + project.  Also helpful to know copyright & author for legal reasons.  For example, if someone was to sell a website design or a company, need to know it is OK to use someone else's code.

Either is not needed, so it is up to you.